Split zng-app-context source for better readability
