+++ crates/zng-app-context/src/lib.rs
@@ -8,7 +8,16 @@
 #![warn(unused_extern_crates)]
 #![warn(missing_docs)]
 
-use std::{any::Any, cell::RefCell, fmt, mem, ops, sync::Arc, thread::LocalKey, time::Duration};
+use std::{any::Any, cell::RefCell, fmt, mem, sync::Arc, thread::LocalKey};
+
+mod util;
+pub use util::*;
+
+mod app_local;
+pub use app_local::*;
+
+mod context_local;
+pub use context_local::*;
 
 use parking_lot::*;
 use zng_txt::Txt;
@@ -446,1100 +455,6 @@ impl LocalKeyDyn for LocalKey<RefCell<LocalData>> {
     }
 }
 
-/*
-    app_local!
-*/
-
-#[doc(hidden)]
-pub struct AppLocalConst<T: Send + Sync + 'static> {
-    value: RwLock<T>,
-}
-impl<T: Send + Sync + 'static> AppLocalConst<T> {
-    pub const fn new(init: T) -> Self {
-        Self { value: RwLock::new(init) }
-    }
-}
-#[doc(hidden)]
-pub struct AppLocalOption<T: Send + Sync + 'static> {
-    value: RwLock<Option<T>>,
-    init: fn() -> T,
-}
-impl<T: Send + Sync + 'static> AppLocalOption<T> {
-    pub const fn new(init: fn() -> T) -> Self {
-        Self {
-            value: RwLock::new(None),
-            init,
-        }
-    }
-
-    fn read_impl(&'static self, read: RwLockReadGuard<'static, Option<T>>) -> MappedRwLockReadGuard<'static, T> {
-        if read.is_some() {
-            return RwLockReadGuard::map(read, |v| v.as_ref().unwrap());
-        }
-        drop(read);
-
-        let mut write = self.value.write();
-        if write.is_some() {
-            drop(write);
-            return self.read();
-        }
-
-        let value = (self.init)();
-        *write = Some(value);
-
-        let read = RwLockWriteGuard::downgrade(write);
-
-        RwLockReadGuard::map(read, |v| v.as_ref().unwrap())
-    }
-
-    fn write_impl(&'static self, mut write: RwLockWriteGuard<'static, Option<T>>) -> MappedRwLockWriteGuard<'static, T> {
-        if write.is_some() {
-            return RwLockWriteGuard::map(write, |v| v.as_mut().unwrap());
-        }
-
-        let value = (self.init)();
-        *write = Some(value);
-
-        RwLockWriteGuard::map(write, |v| v.as_mut().unwrap())
-    }
-}
-
-#[doc(hidden)]
-pub struct AppLocalVec<T: Send + Sync + 'static> {
-    value: RwLock<Vec<(AppId, T)>>,
-    init: fn() -> T,
-}
-impl<T: Send + Sync + 'static> AppLocalVec<T> {
-    pub const fn new(init: fn() -> T) -> Self {
-        Self {
-            value: RwLock::new(vec![]),
-            init,
-        }
-    }
-
-    fn cleanup(&'static self, id: AppId) {
-        self.try_cleanup(id, 0);
-    }
-    fn try_cleanup(&'static self, id: AppId, tries: u8) {
-        if let Some(mut w) = self.value.try_write_for(if tries == 0 {
-            Duration::from_millis(50)
-        } else {
-            Duration::from_millis(500)
-        }) {
-            if let Some(i) = w.iter().position(|(s, _)| *s == id) {
-                w.swap_remove(i);
-            }
-        } else if tries > 5 {
-            tracing::error!("failed to cleanup `app_local` for {id:?}, was locked after app drop");
-        } else {
-            std::thread::spawn(move || {
-                self.try_cleanup(id, tries + 1);
-            });
-        }
-    }
-
-    fn read_impl(&'static self, read: RwLockReadGuard<'static, Vec<(AppId, T)>>) -> MappedRwLockReadGuard<'static, T> {
-        let id = LocalContext::current_app().expect("no app running, `app_local` can only be accessed inside apps");
-
-        if let Some(i) = read.iter().position(|(s, _)| *s == id) {
-            return RwLockReadGuard::map(read, |v| &v[i].1);
-        }
-        drop(read);
-
-        let mut write = self.value.write();
-        if write.iter().any(|(s, _)| *s == id) {
-            drop(write);
-            return self.read();
-        }
-
-        let value = (self.init)();
-        let i = write.len();
-        write.push((id, value));
-
-        LocalContext::register_cleanup(Box::new(move |id| self.cleanup(id)));
-
-        let read = RwLockWriteGuard::downgrade(write);
-
-        RwLockReadGuard::map(read, |v| &v[i].1)
-    }
-
-    fn write_impl(&'static self, mut write: RwLockWriteGuard<'static, Vec<(AppId, T)>>) -> MappedRwLockWriteGuard<'static, T> {
-        let id = LocalContext::current_app().expect("no app running, `app_local` can only be accessed inside apps");
-
-        if let Some(i) = write.iter().position(|(s, _)| *s == id) {
-            return RwLockWriteGuard::map(write, |v| &mut v[i].1);
-        }
-
-        let value = (self.init)();
-        let i = write.len();
-        write.push((id, value));
-
-        LocalContext::register_cleanup(move |id| self.cleanup(id));
-
-        RwLockWriteGuard::map(write, |v| &mut v[i].1)
-    }
-}
-#[doc(hidden)]
-pub trait AppLocalImpl<T: Send + Sync + 'static>: Send + Sync + 'static {
-    fn read(&'static self) -> MappedRwLockReadGuard<'static, T>;
-    fn try_read(&'static self) -> Option<MappedRwLockReadGuard<'static, T>>;
-    fn write(&'static self) -> MappedRwLockWriteGuard<'static, T>;
-    fn try_write(&'static self) -> Option<MappedRwLockWriteGuard<'static, T>>;
-}
-
-impl<T: Send + Sync + 'static> AppLocalImpl<T> for AppLocalVec<T> {
-    fn read(&'static self) -> MappedRwLockReadGuard<'static, T> {
-        self.read_impl(self.value.read_recursive())
-    }
-
-    fn try_read(&'static self) -> Option<MappedRwLockReadGuard<'static, T>> {
-        Some(self.read_impl(self.value.try_read_recursive()?))
-    }
-
-    fn write(&'static self) -> MappedRwLockWriteGuard<'static, T> {
-        self.write_impl(self.value.write())
-    }
-
-    fn try_write(&'static self) -> Option<MappedRwLockWriteGuard<'static, T>> {
-        Some(self.write_impl(self.value.try_write()?))
-    }
-}
-impl<T: Send + Sync + 'static> AppLocalImpl<T> for AppLocalOption<T> {
-    fn read(&'static self) -> MappedRwLockReadGuard<'static, T> {
-        self.read_impl(self.value.read_recursive())
-    }
-
-    fn try_read(&'static self) -> Option<MappedRwLockReadGuard<'static, T>> {
-        Some(self.read_impl(self.value.try_read_recursive()?))
-    }
-
-    fn write(&'static self) -> MappedRwLockWriteGuard<'static, T> {
-        self.write_impl(self.value.write())
-    }
-
-    fn try_write(&'static self) -> Option<MappedRwLockWriteGuard<'static, T>> {
-        Some(self.write_impl(self.value.try_write()?))
-    }
-}
-impl<T: Send + Sync + 'static> AppLocalImpl<T> for AppLocalConst<T> {
-    fn read(&'static self) -> MappedRwLockReadGuard<'static, T> {
-        RwLockReadGuard::map(self.value.read(), |l| l)
-    }
-
-    fn try_read(&'static self) -> Option<MappedRwLockReadGuard<'static, T>> {
-        Some(RwLockReadGuard::map(self.value.try_read()?, |l| l))
-    }
-
-    fn write(&'static self) -> MappedRwLockWriteGuard<'static, T> {
-        RwLockWriteGuard::map(self.value.write(), |l| l)
-    }
-
-    fn try_write(&'static self) -> Option<MappedRwLockWriteGuard<'static, T>> {
-        Some(RwLockWriteGuard::map(self.value.try_write()?, |l| l))
-    }
-}
-
-/// An app local storage.
-///
-/// This is similar to [`std::thread::LocalKey`], but works across all threads of the app.
-///
-/// Use the [`app_local!`] macro to declare a static variable in the same style as [`thread_local!`].
-///
-/// Note that in `"multi_app"` builds the app local can only be used if an app is running in the thread,
-/// if no app is running read and write **will panic**.
-pub struct AppLocal<T: Send + Sync + 'static> {
-    inner: fn() -> &'static dyn AppLocalImpl<T>,
-}
-impl<T: Send + Sync + 'static> AppLocal<T> {
-    #[doc(hidden)]
-    pub const fn new(inner: fn() -> &'static dyn AppLocalImpl<T>) -> Self {
-        AppLocal { inner }
-    }
-
-    /// Read lock the value associated with the current app.
-    ///
-    /// Initializes the default value for the app if this is the first value access.
-    ///
-    /// # Panics
-    ///
-    /// Panics if no app is running in `"multi_app"` builds.
-    #[inline]
-    pub fn read(&'static self) -> MappedRwLockReadGuard<'static, T> {
-        (self.inner)().read()
-    }
-
-    /// Try read lock the value associated with the current app.
-    ///
-    /// Initializes the default value for the app if this is the first value access.
-    ///
-    /// Returns `None` if can’t acquire a read lock.
-    ///
-    /// # Panics
-    ///
-    /// Panics if no app is running in `"multi_app"` builds.
-    #[inline]
-    pub fn try_read(&'static self) -> Option<MappedRwLockReadGuard<'static, T>> {
-        (self.inner)().try_read()
-    }
-
-    /// Write lock the value associated with the current app.
-    ///
-    /// Initializes the default value for the app if this is the first value access.
-    ///
-    /// # Panics
-    ///
-    /// Panics if no app is running in `"multi_app"` builds.
-    #[inline]
-    pub fn write(&'static self) -> MappedRwLockWriteGuard<'static, T> {
-        (self.inner)().write()
-    }
-
-    /// Try to write lock the value associated with the current app.
-    ///
-    /// Initializes the default value for the app if this is the first value access.
-    ///
-    /// Returns `None` if can’t acquire a write lock.
-    ///
-    /// # Panics
-    ///
-    /// Panics if no app is running in `"multi_app"` builds.
-    pub fn try_write(&'static self) -> Option<MappedRwLockWriteGuard<'static, T>> {
-        (self.inner)().try_write()
-    }
-
-    /// Get a clone of the value.
-    #[inline]
-    pub fn get(&'static self) -> T
-    where
-        T: Clone,
-    {
-        self.read().clone()
-    }
-
-    /// Set the value.
-    #[inline]
-    pub fn set(&'static self, value: T) {
-        *self.write() = value;
-    }
-
-    /// Try to get a clone of the value.
-    ///
-    /// Returns `None` if can't acquire a read lock.
-    #[inline]
-    pub fn try_get(&'static self) -> Option<T>
-    where
-        T: Clone,
-    {
-        self.try_read().map(|l| l.clone())
-    }
-
-    /// Try to set the value.
-    ///
-    /// Returns `Err(value)` if can't acquire a write lock.
-    #[inline]
-    pub fn try_set(&'static self, value: T) -> Result<(), T> {
-        match self.try_write() {
-            Some(mut l) => {
-                *l = value;
-                Ok(())
-            }
-            None => Err(value),
-        }
-    }
-
-    /// Create a read lock and `map` it to a sub-value.
-    #[inline]
-    pub fn read_map<O>(&'static self, map: impl FnOnce(&T) -> &O) -> MappedRwLockReadGuard<'static, O> {
-        MappedRwLockReadGuard::map(self.read(), map)
-    }
-
-    /// Try to create a read lock and `map` it to a sub-value.
-    #[inline]
-    pub fn try_read_map<O>(&'static self, map: impl FnOnce(&T) -> &O) -> Option<MappedRwLockReadGuard<'static, O>> {
-        let lock = self.try_read()?;
-        Some(MappedRwLockReadGuard::map(lock, map))
-    }
-
-    /// Create a write lock and `map` it to a sub-value.
-    #[inline]
-    pub fn write_map<O>(&'static self, map: impl FnOnce(&mut T) -> &mut O) -> MappedRwLockWriteGuard<'static, O> {
-        MappedRwLockWriteGuard::map(self.write(), map)
-    }
-
-    /// Try to create a write lock and `map` it to a sub-value.
-    #[inline]
-    pub fn try_write_map<O>(&'static self, map: impl FnOnce(&mut T) -> &mut O) -> Option<MappedRwLockWriteGuard<'static, O>> {
-        let lock = self.try_write()?;
-        Some(MappedRwLockWriteGuard::map(lock, map))
-    }
-
-    /// Gets an ID for this local instance that is valid for the lifetime of the process.
-    ///
-    /// Note that comparing two `&'static LOCAL` pointers is incorrect, because in `"hot_reload"` builds the statics
-    /// can be different and still represent the same app local. This ID identifies the actual inner pointer.
-    pub fn id(&'static self) -> AppLocalId {
-        AppLocalId((self.inner)() as *const dyn AppLocalImpl<T> as *const () as _)
-    }
-}
-impl<T: Send + Sync + 'static> PartialEq for AppLocal<T> {
-    fn eq(&self, other: &Self) -> bool {
-        let a = AppLocalId((self.inner)() as *const dyn AppLocalImpl<T> as *const () as _);
-        let b = AppLocalId((other.inner)() as *const dyn AppLocalImpl<T> as *const () as _);
-        a == b
-    }
-}
-impl<T: Send + Sync + 'static> Eq for AppLocal<T> {}
-impl<T: Send + Sync + 'static> std::hash::Hash for AppLocal<T> {
-    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        let a = AppLocalId((self.inner)() as *const dyn AppLocalImpl<T> as *const () as _);
-        std::hash::Hash::hash(&a, state)
-    }
-}
-
-/// Identifies an [`AppLocal<T>`] instance.
-///
-/// Note that comparing two `&'static LOCAL` pointers is incorrect, because in `"hot_reload"` builds the statics
-/// can be different and still represent the same app local. This ID identifies the actual inner pointer, it is
-/// valid for the lifetime of the process.
-#[derive(PartialEq, Eq, Hash, Clone, Copy)]
-pub struct AppLocalId(usize);
-impl AppLocalId {
-    /// Get the underlying value.
-    pub fn get(self) -> usize {
-        // VarPtr depends on this being an actual pointer (must be unique against an `Arc<T>` raw pointer).
-        self.0 as _
-    }
-}
-impl fmt::Debug for AppLocalId {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "AppLocalId({:#x})", self.0)
-    }
-}
-
-///<span data-del-macro-root></span> Declares new app local variable.
-///
-/// An app local is a static variable that is declared using the same syntax as [`thread_local!`], but can be
-/// accessed by any thread in the app. In apps that only run once per process it compiles down to the equivalent
-/// of a `static LOCAL: RwLock<T> = const;` or `static LOCAL: RwLock<Option<T>>` that initializes on first usage. In test
-/// builds with multiple parallel apps it compiles to a switching storage that provides a different value depending on
-/// what app is running in the current thread.
-///
-/// See [`AppLocal<T>`] for more details.
-///
-/// # Multi App
-///
-/// If the crate is compiled with the `"multi_app"` feature a different internal implementation is used that supports multiple
-/// apps, either running in parallel in different threads or one after the other. This backing implementation has some small overhead,
-/// but usually you only want multiple app instances per-process when running tests.
-///
-/// The lifetime of `"multi_app"` locals is also more limited, trying to use an app-local before starting to build an app will panic,
-/// the app-local value will be dropped when the app is dropped. Without the `"multi_app"` feature the app-locals can be used at
-/// any point before or after the app lifetime, values are not explicitly dropped, just unloaded with the process.
-///
-/// # Const
-///
-/// The initialization expression can be wrapped in a `const { .. }` block, if the `"multi_app"` feature is **not** enabled
-/// a faster implementation is used that is equivalent to a direct `static LOCAL: RwLock<T>` in terms of performance.
-///
-/// Note that this syntax is available even if the `"multi_app"` feature is enabled, the expression must be const either way,
-/// but with the feature the same dynamic implementation is used.
-///
-/// Note that `const` initialization does not automatically convert the value into the static type.
-///
-/// # Examples
-///
-/// The example below declares two app locals, note that `BAR` init value automatically converts into the app local type.
-///
-/// ```
-/// # use zng_app_context::*;
-/// app_local! {
-///     /// A public documented value.
-///     pub static FOO: u8 = const { 10u8 };
-///
-///     // A private value.
-///     static BAR: String = "Into!";
-/// }
-///
-/// let app = LocalContext::start_app(AppId::new_unique());
-///
-/// assert_eq!(10, FOO.get());
-/// ```
-///
-/// Also note that an app context is started before the first use, in `multi_app` builds trying to use an app local in
-/// a thread not owned by an app panics.
-#[macro_export]
-macro_rules! app_local {
-    ($(
-        $(#[$meta:meta])*
-        $vis:vis static $IDENT:ident : $T:ty = $(const { $init_const:expr })? $($init:expr_2021)?;
-    )+) => {$(
-        $crate::app_local_impl! {
-            $(#[$meta])*
-            $vis static $IDENT: $T = $(const { $init_const })? $($init)?;
-        }
-    )+};
-}
-
-#[doc(hidden)]
-#[macro_export]
-macro_rules! app_local_impl_single {
-    (
-        $(#[$meta:meta])*
-        $vis:vis static $IDENT:ident : $T:ty = const { $init:expr };
-    ) => {
-        $(#[$meta])*
-        $vis static $IDENT: $crate::AppLocal<$T> = {
-            fn s() -> &'static dyn $crate::AppLocalImpl<$T> {
-                $crate::hot_static! {
-                    static IMPL: $crate::AppLocalConst<$T> = $crate::AppLocalConst::new($init);
-                }
-                $crate::hot_static_ref!(IMPL)
-            }
-            $crate::AppLocal::new(s)
-        };
-    };
-    (
-        $(#[$meta:meta])*
-        $vis:vis static $IDENT:ident : $T:ty = $init:expr_2021;
-    ) => {
-        $(#[$meta])*
-        $vis static $IDENT: $crate::AppLocal<$T> = {
-            fn s() -> &'static dyn $crate::AppLocalImpl<$T> {
-                fn init() -> $T {
-                    std::convert::Into::into($init)
-                }
-                $crate::hot_static! {
-                    static IMPL: $crate::AppLocalOption<$T> = $crate::AppLocalOption::new(init);
-                }
-                $crate::hot_static_ref!(IMPL)
-            }
-            $crate::AppLocal::new(s)
-        };
-    };
-    (
-        $(#[$meta:meta])*
-        $vis:vis static $IDENT:ident : $T:ty = ($tt:tt)*
-    ) => {
-        std::compile_error!("expected `const { $expr };` or `$expr;`")
-    };
-}
-
-#[doc(hidden)]
-#[macro_export]
-macro_rules! app_local_impl_multi {
-    (
-        $(#[$meta:meta])*
-        $vis:vis static $IDENT:ident : $T:ty = const { $init:expr };
-    ) => {
-        $(#[$meta])*
-        $vis static $IDENT: $crate::AppLocal<$T> = {
-            fn s() -> &'static dyn $crate::AppLocalImpl<$T> {
-                const fn init() -> $T {
-                    $init
-                }
-                $crate::hot_static! {
-                    static IMPL: $crate::AppLocalVec<$T> = $crate::AppLocalVec::new(init);
-                }
-                $crate::hot_static_ref!(IMPL)
-            }
-            $crate::AppLocal::new(s)
-        };
-    };
-    (
-        $(#[$meta:meta])*
-        $vis:vis static $IDENT:ident : $T:ty = $init:expr_2021;
-    ) => {
-        $(#[$meta])*
-        $vis static $IDENT: $crate::AppLocal<$T> = {
-            fn s() -> &'static dyn $crate::AppLocalImpl<$T> {
-                fn init() -> $T {
-                    std::convert::Into::into($init)
-                }
-                $crate::hot_static! {
-                    static IMPL: $crate::AppLocalVec<$T> = $crate::AppLocalVec::new(init);
-                }
-                $crate::hot_static_ref!(IMPL)
-            }
-            $crate::AppLocal::new(s)
-        };
-    };
-    (
-        $(#[$meta:meta])*
-        $vis:vis static $IDENT:ident : $T:ty = ($tt:tt)*
-    ) => {
-        std::compile_error!("expected `const { $expr };` or `$expr;`")
-    };
-}
-
-#[cfg(feature = "multi_app")]
-#[doc(hidden)]
-pub use app_local_impl_multi as app_local_impl;
-#[cfg(not(feature = "multi_app"))]
-#[doc(hidden)]
-pub use app_local_impl_single as app_local_impl;
-
-/*
-    context_local!
-*/
-
-#[doc(hidden)]
-pub struct ContextLocalData<T: Send + Sync + 'static> {
-    default_init: fn() -> T,
-    default_value: Option<Arc<T>>,
-}
-impl<T: Send + Sync + 'static> ContextLocalData<T> {
-    #[doc(hidden)]
-    pub const fn new(default_init: fn() -> T) -> Self {
-        Self {
-            default_init,
-            default_value: None,
-        }
-    }
-}
-
-/// Represents an [`AppLocal<T>`] value that can be temporarily overridden in a context.
-///
-/// The *context* works across threads, as long as the threads are instrumented using [`LocalContext`].
-///
-/// Use the [`context_local!`] macro to declare a static variable in the same style as [`thread_local!`].
-pub struct ContextLocal<T: Send + Sync + 'static> {
-    data: AppLocal<ContextLocalData<T>>,
-}
-impl<T: Send + Sync + 'static> ContextLocal<T> {
-    #[doc(hidden)]
-    pub const fn new(storage: fn() -> &'static dyn AppLocalImpl<ContextLocalData<T>>) -> Self {
-        Self {
-            data: AppLocal::new(storage),
-        }
-    }
-
-    /// Gets an ID for this context local instance that is valid for the lifetime of the process.
-    ///
-    /// Note that comparing two `&'static CTX_LOCAL` pointers is incorrect, because in `"hot_reload"` builds the statics
-    /// can be different and still represent the same app local. This ID identifies the actual inner pointer.
-    pub fn id(&'static self) -> AppLocalId {
-        self.data.id()
-    }
-
-    /// Calls `f` with the `value` loaded in context.
-    ///
-    /// The `value` is moved into context, `f` is called, then the value is moved back to `value`.
-    ///
-    /// # Panics
-    ///
-    /// Panics if `value` is `None`.
-    pub fn with_context<R>(&'static self, value: &mut Option<Arc<T>>, f: impl FnOnce() -> R) -> R {
-        let mut r = None;
-        let f = || r = Some(f());
-        #[cfg(feature = "dyn_closure")]
-        let f: Box<dyn FnOnce()> = Box::new(f);
-
-        LocalContext::with_value_ctx(self, LocalValueKind::Local, value, f);
-
-        r.unwrap()
-    }
-
-    /// Same as [`with_context`], but `value` represents a variable.
-    ///
-    /// Values loaded with this method are captured by [`CaptureFilter::ContextVars`].
-    ///
-    /// [`with_context`]: Self::with_context
-    pub fn with_context_var<R>(&'static self, value: &mut Option<Arc<T>>, f: impl FnOnce() -> R) -> R {
-        let mut r = None;
-        let f = || r = Some(f());
-
-        #[cfg(feature = "dyn_closure")]
-        let f: Box<dyn FnOnce()> = Box::new(f);
-
-        LocalContext::with_value_ctx(self, LocalValueKind::Var, value, f);
-
-        r.unwrap()
-    }
-
-    /// Calls `f` with no value loaded in context.
-    pub fn with_default<R>(&'static self, f: impl FnOnce() -> R) -> R {
-        let mut r = None;
-        let f = || r = Some(f());
-
-        #[cfg(feature = "dyn_closure")]
-        let f: Box<dyn FnOnce()> = Box::new(f);
-        LocalContext::with_default_ctx(self, f);
-
-        r.unwrap()
-    }
-
-    /// Gets if no value is set in the context.
-    pub fn is_default(&'static self) -> bool {
-        !LocalContext::contains(self.id())
-    }
-
-    /// Clone a reference to the current value in the context or the default value.
-    pub fn get(&'static self) -> Arc<T> {
-        let cl = self.data.read();
-        match LocalContext::get(self.id()) {
-            Some(c) => Arc::downcast(c.0).unwrap(),
-            None => match &cl.default_value {
-                Some(d) => d.clone(),
-                None => {
-                    drop(cl);
-                    let mut cl = self.data.write();
-                    match &cl.default_value {
-                        None => {
-                            let d = Arc::new((cl.default_init)());
-                            cl.default_value = Some(d.clone());
-                            d
-                        }
-                        Some(d) => d.clone(),
-                    }
-                }
-            },
-        }
-    }
-
-    /// Clone the current value in the context or the default value.
-    pub fn get_clone(&'static self) -> T
-    where
-        T: Clone,
-    {
-        let cl = self.data.read();
-        match LocalContext::get(self.id()) {
-            Some(c) => c.0.downcast_ref::<T>().unwrap().clone(),
-            None => match &cl.default_value {
-                Some(d) => d.as_ref().clone(),
-                None => {
-                    drop(cl);
-                    let mut cl = self.data.write();
-                    match &cl.default_value {
-                        None => {
-                            let val = (cl.default_init)();
-                            let r = val.clone();
-                            cl.default_value = Some(Arc::new(val));
-                            r
-                        }
-                        Some(d) => d.as_ref().clone(),
-                    }
-                }
-            },
-        }
-    }
-}
-
-impl<T: Send + Sync + 'static> ContextLocal<RwLock<T>> {
-    /// Gets a read-only shared reference to the current context value.
-    pub fn read_only(&'static self) -> ReadOnlyRwLock<T> {
-        ReadOnlyRwLock::new(self.get())
-    }
-
-    /// Locks this `RwLock` with shared read access, blocking the current thread until it can be acquired.
-    ///
-    /// See `parking_lot::RwLock::read` for more details.
-    pub fn read(&'static self) -> RwLockReadGuardOwned<T> {
-        RwLockReadGuardOwned::lock(self.get())
-    }
-
-    /// Locks this `RwLock` with shared read access, blocking the current thread until it can be acquired.
-    ///
-    /// Unlike `read`, this method is guaranteed to succeed without blocking if
-    /// another read lock is held at the time of the call.
-    ///
-    /// See `parking_lot::RwLock::read` for more details.
-    pub fn read_recursive(&'static self) -> RwLockReadGuardOwned<T> {
-        RwLockReadGuardOwned::lock_recursive(self.get())
-    }
-
-    /// Locks this `RwLock` with exclusive write access, blocking the current
-    /// thread until it can be acquired.
-    ///
-    /// See `parking_lot::RwLock::write` for more details.
-    pub fn write(&'static self) -> RwLockWriteGuardOwned<T> {
-        RwLockWriteGuardOwned::lock(self.get())
-    }
-
-    /// Try lock this `RwLock` with shared read access, blocking the current thread until it can be acquired.
-    ///
-    /// See `parking_lot::RwLock::try_read` for more details.
-    pub fn try_read(&'static self) -> Option<RwLockReadGuardOwned<T>> {
-        RwLockReadGuardOwned::try_lock(self.get())
-    }
-
-    /// Locks this `RwLock` with shared read access, blocking the current thread until it can be acquired.
-    ///
-    /// See `parking_lot::RwLock::try_read_recursive` for more details.
-    pub fn try_read_recursive(&'static self) -> Option<RwLockReadGuardOwned<T>> {
-        RwLockReadGuardOwned::try_lock_recursive(self.get())
-    }
-
-    /// Locks this `RwLock` with exclusive write access, blocking the current
-    /// thread until it can be acquired.
-    ///
-    /// See `parking_lot::RwLock::try_write` for more details.
-    pub fn try_write(&'static self) -> Option<RwLockWriteGuardOwned<T>> {
-        RwLockWriteGuardOwned::try_lock(self.get())
-    }
-}
-
-/// Represents a read guard for an `Arc<RwLock<T>>` that owns a reference to it.
-pub struct RwLockReadGuardOwned<T: 'static> {
-    lock: parking_lot::RwLockReadGuard<'static, T>,
-    _owned: Arc<RwLock<T>>,
-}
-impl<T> RwLockReadGuardOwned<T> {
-    /// Lock owned.    
-    ///
-    /// See `parking_lot::RwLock::read` for more details.
-    pub fn lock(own: Arc<RwLock<T>>) -> Self {
-        Self {
-            // SAFETY: we cast to 'static only for storage, `lock` is dropped before `_owned`.
-            lock: unsafe { mem::transmute::<parking_lot::RwLockReadGuard<'_, T>, parking_lot::RwLockReadGuard<'static, T>>(own.read()) },
-            _owned: own,
-        }
-    }
-
-    /// Locks this `RwLock` with shared read access, blocking the current thread until it can be acquired.
-    ///
-    /// See `parking_lot::RwLock::read_recursive` for more details.
-    pub fn lock_recursive(own: Arc<RwLock<T>>) -> Self {
-        Self {
-            // SAFETY: we cast to 'static only for storage, `lock` is dropped before `_owned`.
-            lock: unsafe {
-                mem::transmute::<parking_lot::RwLockReadGuard<'_, T>, parking_lot::RwLockReadGuard<'static, T>>(own.read_recursive())
-            },
-            _owned: own,
-        }
-    }
-
-    /// Try lock owned.
-    ///
-    /// See `parking_lot::RwLock::try_read` for more details.
-    pub fn try_lock(own: Arc<RwLock<T>>) -> Option<Self> {
-        let lock = own.try_read()?;
-        Some(Self {
-            // SAFETY: we cast to 'static only for storage, `lock` is dropped before `_owned`.
-            lock: unsafe { mem::transmute::<parking_lot::RwLockReadGuard<'_, T>, parking_lot::RwLockReadGuard<'static, T>>(lock) },
-            _owned: own,
-        })
-    }
-
-    /// Try lock owned.
-    ///
-    /// See `parking_lot::RwLock::try_read` for more details.
-    pub fn try_lock_recursive(own: Arc<RwLock<T>>) -> Option<Self> {
-        let lock = own.try_read_recursive()?;
-        Some(Self {
-            // SAFETY: we cast to 'static only for storage, `lock` is dropped before `_owned`.
-            lock: unsafe { mem::transmute::<parking_lot::RwLockReadGuard<'_, T>, parking_lot::RwLockReadGuard<'static, T>>(lock) },
-            _owned: own,
-        })
-    }
-
-    /// Make a new `MappedRwLockReadGuardOwned` for a component of the locked data.
-    ///
-    /// This is an associated function that needs to be
-    /// used as `RwLockReadGuardOwned::map(...)`. A method would interfere with methods of
-    /// the same name on the contents of the locked data.
-    pub fn map<O>(guard: Self, map: impl FnOnce(&T) -> &O) -> MappedRwLockReadGuardOwned<T, O> {
-        MappedRwLockReadGuardOwned {
-            lock: parking_lot::RwLockReadGuard::map(guard.lock, map),
-            _owned: guard._owned,
-        }
-    }
-}
-impl<T> ops::Deref for RwLockReadGuardOwned<T> {
-    type Target = T;
-
-    fn deref(&self) -> &Self::Target {
-        self.lock.deref()
-    }
-}
-
-/// Represents a read guard for an `Arc<RwLock<T>>` that owns a reference to it, mapped from another read guard.
-pub struct MappedRwLockReadGuardOwned<T: 'static, O: 'static> {
-    lock: parking_lot::MappedRwLockReadGuard<'static, O>,
-    _owned: Arc<RwLock<T>>,
-}
-impl<T, O> MappedRwLockReadGuardOwned<T, O> {
-    /// Make a new `MappedRwLockReadGuardOwned` for a component of the locked data.
-    ///
-    /// This is an associated function that needs to be
-    /// used as `MappedRwLockReadGuardOwned::map(...)`. A method would interfere with methods of
-    /// the same name on the contents of the locked data.
-    pub fn map<O2>(guard: Self, map: impl FnOnce(&O) -> &O2) -> MappedRwLockReadGuardOwned<T, O2> {
-        MappedRwLockReadGuardOwned {
-            lock: parking_lot::MappedRwLockReadGuard::map(guard.lock, map),
-            _owned: guard._owned,
-        }
-    }
-}
-impl<T, O> ops::Deref for MappedRwLockReadGuardOwned<T, O> {
-    type Target = O;
-
-    fn deref(&self) -> &Self::Target {
-        self.lock.deref()
-    }
-}
-
-/// Represents a read guard for an `Arc<RwLock<T>>` that owns a reference to it.
-pub struct RwLockWriteGuardOwned<T: 'static> {
-    lock: parking_lot::RwLockWriteGuard<'static, T>,
-    _owned: Arc<RwLock<T>>,
-}
-impl<T> RwLockWriteGuardOwned<T> {
-    /// Lock owned.
-    ///
-    /// See `parking_lot::RwLock::write` for more details.
-    pub fn lock(own: Arc<RwLock<T>>) -> Self {
-        Self {
-            // SAFETY: we cast to 'static only for storage, `lock` is dropped before `_owned`.
-            lock: unsafe { mem::transmute::<parking_lot::RwLockWriteGuard<'_, T>, parking_lot::RwLockWriteGuard<'static, T>>(own.write()) },
-            _owned: own,
-        }
-    }
-
-    /// Lock owned.
-    ///
-    /// See `parking_lot::RwLock::try_write` for more details.
-    pub fn try_lock(own: Arc<RwLock<T>>) -> Option<Self> {
-        let lock = own.try_write()?;
-        Some(Self {
-            // SAFETY: we cast to 'static only for storage, `lock` is dropped before `_owned`.
-            lock: unsafe { mem::transmute::<parking_lot::RwLockWriteGuard<'_, T>, parking_lot::RwLockWriteGuard<'static, T>>(lock) },
-            _owned: own,
-        })
-    }
-
-    /// Make a new `MappedRwLockReadGuardOwned` for a component of the locked data.
-    ///
-    /// This is an associated function that needs to be
-    /// used as `MappedRwLockReadGuardOwned::map(...)`. A method would interfere with methods of
-    /// the same name on the contents of the locked data.
-    pub fn map<O>(guard: Self, map: impl FnOnce(&mut T) -> &mut O) -> MappedRwLockWriteGuardOwned<T, O> {
-        MappedRwLockWriteGuardOwned {
-            lock: parking_lot::RwLockWriteGuard::map(guard.lock, map),
-            _owned: guard._owned,
-        }
-    }
-}
-impl<T> ops::Deref for RwLockWriteGuardOwned<T> {
-    type Target = T;
-
-    fn deref(&self) -> &Self::Target {
-        self.lock.deref()
-    }
-}
-impl<T> ops::DerefMut for RwLockWriteGuardOwned<T> {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        self.lock.deref_mut()
-    }
-}
-
-/// Represents a write guard for an `Arc<RwLock<T>>` that owns a reference to it, mapped from another read guard.
-pub struct MappedRwLockWriteGuardOwned<T: 'static, O: 'static> {
-    lock: parking_lot::MappedRwLockWriteGuard<'static, O>,
-    _owned: Arc<RwLock<T>>,
-}
-impl<T, O> MappedRwLockWriteGuardOwned<T, O> {
-    /// Make a new `MappedRwLockWriteGuardOwned` for a component of the locked data.
-    ///
-    /// This is an associated function that needs to be
-    /// used as `MappedRwLockWriteGuardOwned::map(...)`. A method would interfere with methods of
-    /// the same name on the contents of the locked data.
-    pub fn map<O2>(guard: Self, map: impl FnOnce(&mut O) -> &mut O2) -> MappedRwLockWriteGuardOwned<T, O2> {
-        MappedRwLockWriteGuardOwned {
-            lock: parking_lot::MappedRwLockWriteGuard::map(guard.lock, map),
-            _owned: guard._owned,
-        }
-    }
-}
-impl<T, O> ops::Deref for MappedRwLockWriteGuardOwned<T, O> {
-    type Target = O;
-
-    fn deref(&self) -> &Self::Target {
-        self.lock.deref()
-    }
-}
-impl<T, O> ops::DerefMut for MappedRwLockWriteGuardOwned<T, O> {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        self.lock.deref_mut()
-    }
-}
-
-/// Read-only wrapper on an `Arc<RwLock<T>>` contextual value.
-pub struct ReadOnlyRwLock<T>(Arc<RwLock<T>>);
-impl<T> Clone for ReadOnlyRwLock<T> {
-    fn clone(&self) -> Self {
-        Self(self.0.clone())
-    }
-}
-impl<T> ReadOnlyRwLock<T> {
-    /// New.
-    pub fn new(l: Arc<RwLock<T>>) -> Self {
-        Self(l)
-    }
-
-    /// Locks this `RwLock` with shared read access, blocking the current thread until it can be acquired.
-    ///
-    /// See `parking_lot::RwLock::read` for more details.
-    pub fn read(&self) -> parking_lot::RwLockReadGuard<'_, T> {
-        self.0.read()
-    }
-
-    /// Locks this `RwLock` with shared read access, blocking the current thread until it can be acquired.
-    ///
-    /// Unlike `read`, this method is guaranteed to succeed without blocking if
-    /// another read lock is held at the time of the call.
-    ///
-    /// See `parking_lot::RwLock::read_recursive` for more details.
-    pub fn read_recursive(&self) -> parking_lot::RwLockReadGuard<'_, T> {
-        self.0.read_recursive()
-    }
-
-    /// Attempts to acquire this `RwLock` with shared read access.
-    ///
-    /// See `parking_lot::RwLock::try_read` for more details.
-    pub fn try_read(&self) -> Option<parking_lot::RwLockReadGuard<'_, T>> {
-        self.0.try_read()
-    }
-
-    /// Attempts to acquire this `RwLock` with shared read access.
-    ///
-    /// See `parking_lot::RwLock::try_read_recursive` for more details.
-    pub fn try_read_recursive(&self) -> Option<parking_lot::RwLockReadGuard<'_, T>> {
-        self.0.try_read_recursive()
-    }
-
-    /// Gets if the read-only shared reference is to the same lock as `other`.
-    pub fn ptr_eq(&self, other: &Self) -> bool {
-        Arc::ptr_eq(&self.0, &other.0)
-    }
-}
-
-///<span data-del-macro-root></span> Declares new app and context local variable.
-///
-/// # Examples
-///
-/// ```
-/// # use zng_app_context::*;
-/// context_local! {
-///     /// A public documented value.
-///     pub static FOO: u8 = 10u8;
-///
-///     // A private value.
-///     static BAR: String = "Into!";
-/// }
-/// ```
-///
-/// # Default Value
-///
-/// All contextual values must have a fallback value that is used when no context is loaded.
-///
-/// The default value is instantiated once per app, the expression can be any static value that converts [`Into<T>`].
-///
-/// # Usage
-///
-/// After you declare the context local you can use it by loading a contextual value for the duration of a closure call.
-///
-/// ```
-/// # use zng_app_context::*;
-/// # use std::sync::Arc;
-/// context_local! { static FOO: String = "default"; }
-///
-/// fn print_value() {
-///     println!("value is {}!", FOO.get());
-/// }
-///
-/// let _scope = LocalContext::start_app(AppId::new_unique());
-///
-/// let mut value = Some(Arc::new(String::from("other")));
-/// FOO.with_context(&mut value, || {
-///     print!("in context, ");
-///     print_value();
-/// });
-///
-/// print!("out of context, ");
-/// print_value();
-/// ```
-///
-/// The example above prints:
-///
-/// ```text
-/// in context, value is other!
-/// out of context, value is default!
-/// ```
-///
-/// See [`ContextLocal<T>`] for more details.
-#[macro_export]
-macro_rules! context_local {
-    ($(
-        $(#[$meta:meta])*
-        $vis:vis static $IDENT:ident : $T:ty = $init:expr;
-    )+) => {$(
-        $crate::context_local_impl! {
-            $(#[$meta])*
-            $vis static $IDENT: $T = $init;
-        }
-    )+};
-}
-
-#[doc(hidden)]
-#[macro_export]
-macro_rules! context_local_impl_single {
-    ($(
-        $(#[$meta:meta])*
-        $vis:vis static $IDENT:ident : $T:ty = $init:expr;
-    )+) => {$(
-        $(#[$meta])*
-        $vis static $IDENT: $crate::ContextLocal<$T> = {
-            fn s() -> &'static dyn $crate::AppLocalImpl<$crate::ContextLocalData<$T>> {
-                fn init() -> $T {
-                    std::convert::Into::into($init)
-                }
-                $crate::hot_static! {
-                    static IMPL: $crate::AppLocalConst<$crate::ContextLocalData<$T>> =
-                    $crate::AppLocalConst::new(
-                        $crate::ContextLocalData::new(init)
-                    );
-                }
-                $crate::hot_static_ref!(IMPL)
-            }
-            $crate::ContextLocal::new(s)
-        };
-    )+};
-}
-
-#[doc(hidden)]
-#[macro_export]
-macro_rules! context_local_impl_multi {
-    ($(
-        $(#[$meta:meta])*
-        $vis:vis static $IDENT:ident : $T:ty = $init:expr;
-    )+) => {$(
-        $(#[$meta])*
-        $vis static $IDENT: $crate::ContextLocal<$T> = {
-            fn s() -> &'static dyn $crate::AppLocalImpl<$crate::ContextLocalData<$T>> {
-                fn init() -> $T {
-                    std::convert::Into::into($init)
-                }
-                $crate::hot_static! {
-                    static IMPL: $crate::AppLocalVec<$crate::ContextLocalData<$T>> =
-                    $crate::AppLocalVec::new(
-                        || $crate::ContextLocalData::new(init)
-                    );
-                }
-                $crate::hot_static_ref!(IMPL)
-            }
-            $crate::ContextLocal::new(s)
-        };
-    )+};
-}
-
-#[cfg(feature = "multi_app")]
-#[doc(hidden)]
-pub use context_local_impl_multi as context_local_impl;
-
-#[cfg(not(feature = "multi_app"))]
-#[doc(hidden)]
-pub use context_local_impl_single as context_local_impl;
-
 /// Defines a [`LocalContext::capture_filtered`] filter.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum CaptureFilter {
@@ -1670,31 +585,5 @@ impl ContextValueSet {
 impl fmt::Debug for ContextValueSet {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.debug_struct("ContextValueSet").field("len()", &self.len()).finish()
-    }
-}
-
-/// Helper, runs a cleanup action once on drop.
-pub struct RunOnDrop<F: FnOnce()>(Option<F>);
-impl<F: FnOnce()> RunOnDrop<F> {
-    /// New with closure that will run once on drop.
-    pub fn new(clean: F) -> Self {
-        RunOnDrop(Some(clean))
-    }
-}
-impl<F: FnOnce()> Drop for RunOnDrop<F> {
-    fn drop(&mut self) {
-        if let Some(clean) = self.0.take() {
-            clean();
-        }
-    }
-}
-
-fn panic_str<'s>(payload: &'s Box<dyn std::any::Any + Send + 'static>) -> &'s str {
-    if let Some(s) = payload.downcast_ref::<&str>() {
-        s
-    } else if let Some(s) = payload.downcast_ref::<String>() {
-        s
-    } else {
-        "<unknown-panic-message-type>"
     }
 }
